Add Report Design branding, Learner Auto-Create, negative marking, Reported Questions, and theming
Bundles the session's accumulated feature work into one push:
- Report Design settings: PDF template/accent-color picker plus business
  name, logo upload, and address, threaded through every PDF exporter's
  letterhead.
- Learners: Auto Create toggle that creates a learner profile via a
  security-definer RPC when a quiz is submitted through the public
  share-link flow, plus Profile Details summary stats and PDF/CSV export
  with a date-range picker.
- Negative marking: decimal question points (Int -> Double), a +/-
  stepper replacing the points slider, and per-question negative scoring
  that floors the total at 0%.
- Reported Questions screen plus Dashboard stat tiles for reported
  questions and learner count.
- Onboarding intro screens for Classes/Learners/Tools, light/dark theme
  support, and scheduled push notifications.
- Sale Day pricing now auto-applies the discount and countdown animation
  the moment a live sale is in view, instead of requiring a manual
  "Unlock" tap.
- Fixes: stale previous-account data surviving logout, PostHog identify
  not firing on a restored session, and phone-number-entry country-code/
  field misalignment on the Collect Phone screen.

+++ app/src/main/java/com/quizmaker/android/data/model/Quiz.kt
@@ -28,7 +28,7 @@ data class Quiz(
     val collectAddress: Boolean,
     val collectPhone: Boolean,
     val quizColor: String,
-    val maxPoints: Int,
+    val maxPoints: Double,
     val createdAt: Instant?
 ) {
     val shareUrl: String get() = "${BuildConfig.SHARE_BASE_URL}/take-quiz/$shareId"
@@ -79,6 +79,6 @@ fun QuizDto.toDomain(): Quiz = Quiz(
     collectAddress = collectAddress ?: false,
     collectPhone = collectPhone ?: false,
     quizColor = quizColor ?: "#8b5cf6",
-    maxPoints = maxPoints ?: 0,
+    maxPoints = maxPoints ?: 0.0,
     createdAt = createdAt?.let { runCatching { Instant.parse(it) }.getOrNull() }
 )
